# Reduced multi_fault.BLOCKSIZE