--- a/openquake/hazardlib/source/multi_fault.py
+++ b/openquake/hazardlib/source/multi_fault.py
@@ -31,7 +31,7 @@ from openquake.hazardlib.geo.utils import angular_distance, KM_TO_DEGREES
 from openquake.hazardlib.source.base import BaseSeismicSource
 
 F32 = np.float32
-BLOCKSIZE = 1000
+BLOCKSIZE = 100
 # the BLOCKSIZE has to be large to reduce the number of sources and
 # therefore the redundant data transfer in the .sections attribute